Code improvements
ESLint picked up a few points of improvement.
Split the functionality of the temporary email table to a seperated
function that's triggered by Firestore. This will also help if we ever
manage to migrate the old data into Firebase.
Added functionality to prepare the tables that will be used by the
"Trigger Email" extension.