Rewrite notes test to be async
During the test it would sometimes create a fully paid document first
and then it would overwrite it with an initial payment document.
Thus, now we wait for the initial payment to be created.
Since we now have a lot of sleeps, also increase the
timeout from mocha to 10s.